# apps/fe/src/shared/api/openapi/api-1.yaml
openapi: 3.0.0
paths:
...
  /api/students/{studentId}/external-accounts:
    get:
      operationId: ExternalAccountsController_list
      parameters:
        - name: studentId
          required: true
          in: path
          schema:
            format: uuid
            example: 7f4a8c2d-1b7e-4a7f-9d3b-2c6f4e8a9b10
            type: string
      responses:
        "200":
          description: ""
      security: &a4
        - bearer: []
      tags: &a5
        - External Accounts
    post:
      operationId: ExternalAccountsController_create
      parameters:
        - name: studentId
          required: true
          in: path
          schema:
            format: uuid
            example: 7f4a8c2d-1b7e-4a7f-9d3b-2c6f4e8a9b10
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CreateExternalAccountDto"
              example:
                platform: LICHESS
                username: maksim_sokolov_2012
      responses:
        "201":
          description: ""
      security: *a4
      tags: *a5
  /api/students/{studentId}/external-accounts/{externalAccountId}:
    patch:
      operationId: ExternalAccountsController_update
      parameters:
        - name: studentId
          required: true
          in: path
          schema:
            format: uuid
            example: 7f4a8c2d-1b7e-4a7f-9d3b-2c6f4e8a9b10
            type: string
        - name: externalAccountId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExternalAccountDto"
      responses:
        "200":
          description: ""
      security: *a4
      tags: *a5
    delete:
      operationId: ExternalAccountsController_remove
      parameters:
        - name: studentId
          required: true
          in: path
          schema:
            format: uuid
            example: 7f4a8c2d-1b7e-4a7f-9d3b-2c6f4e8a9b10
            type: string
        - name: externalAccountId
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: ""
      security: *a4
      tags: *a5
  /api/students/{studentId}/games:
    get:
      operationId: GamesController_listStudentGames
      parameters:
        - name: studentId
          required: true
          in: path
          schema:
            type: string
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 50
            default: 20
            type: number
        - name: cursor
          required: false
          in: query
          schema:
            format: uuid
            type: string
        - name: analysisStatus
          required: false
          in: query
          schema:
            type: string
            enum:
              - PENDING
              - PARSING
              - EXTRACTING_ANNOTATIONS
              - CLASSIFICATION
              - GENERATING_OUTPUT
              - COMPLETED
              - FAILED
      responses:
        "200":
          description: ""
      security: &a6
        - bearer: []
      tags: &a7
        - Games
  /api/games/{gameId}:
    get:
      operationId: GamesController_getOne
      parameters:
        - name: gameId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a6
      tags: *a7
  /api/games/{gameId}/pgn:
    get:
      operationId: GamesController_getPgn
      parameters:
        - name: gameId
          required: true
          in: path
          schema:
            type: string
        - name: download
          required: false
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: ""
      security: *a6
      tags: *a7
  /api/analysis/jobs:
    get:
      operationId: AnalysisJobsController_list
      parameters:
        - name: studentId
          required: false
          in: query
          schema:
            format: uuid
            type: string
        - name: gameId
          required: false
          in: query
          schema:
            format: uuid
            type: string
        - name: jobType
          required: false
          in: query
          schema:
            type: string
            enum:
              - ANALYSIS
              - REPORT_GENERATION
              - HOMEWORK_GENERATION
              - PROGRESS_GENERATION
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - PENDING
              - PARSING
              - EXTRACTING_ANNOTATIONS
              - CLASSIFICATION
              - GENERATING_OUTPUT
              - COMPLETED
              - FAILED
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 50
            default: 20
            type: number
        - name: cursor
          required: false
          in: query
          schema:
            format: uuid
            type: string
      responses:
        "200":
          description: ""
      security: &a8
        - bearer: []
      tags: &a9
        - Analysis Jobs
  /api/analysis/jobs/{jobId}:
    get:
      operationId: AnalysisJobsController_getStatus
      parameters:
        - name: jobId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a8
      tags: *a9
  /api/analysis/jobs/{jobId}/retry:
    post:
      operationId: AnalysisJobsController_retry
      parameters:
        - name: jobId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a8
      tags: *a9
  /api/analysis:
    get:
      operationId: AnalysisController_list
      parameters:
        - name: studentId
          required: false
          in: query
          schema:
            format: uuid
            type: string
      responses:
        "200":
          description: ""
      security: &a10
        - bearer: []
      tags: &a11
        - Analysis
  /api/analysis/{analysisId}:
    get:
      operationId: AnalysisController_getOne
      parameters:
        - name: analysisId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/AnalysisDetailsResponse"
      security: *a10
      tags: *a11
  /api/students/{studentId}/imports/pgn:
    post:
      operationId: ImportsController_importPgn
      parameters:
        - name: studentId
          required: true
          in: path
          schema:
            format: uuid
            example: 7f4a8c2d-1b7e-4a7f-9d3b-2c6f4e8a9b10
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/ImportPgnDto"
              example:
                studentColor: WHITE
                sourceLabel: Lichess Study export
                rawPgn: >-
                  [Event "Training Game"]

                  [Site "Lichess"]

                  [Date "2026.07.25"]

                  [Round "?"]

                  [White "Student"]

                  [Black "Opponent"]

                  [Result "1-0"]


                  1. e4 { [%eval 0.3] Good central control. } e5 { [%eval 0.1] }
                  2. Nf3 { [%eval 0.5] } Nc6 { [%eval 0.2] } 3. Bb5 { [%eval
                  0.9] } a6 { [%eval 0.4] } 1-0
      responses:
        "201":
          description: ""
      security:
        - bearer: []
      tags:
        - Imports
...
  /api/analysis/{analysisId}/homework/generate:
    post:
      operationId: HomeworkController_requestGeneration
      parameters:
        - name: analysisId
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      security: &a14
        - bearer: []
      tags: &a15
        - Homework
  /api/homework:
    get:
      operationId: HomeworkController_list
      parameters:
        - name: studentId
          required: true
          in: query
          schema:
            type: string
        - name: analysisId
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a14
      tags: *a15
  /api/homework/{homeworkId}:
    get:
      operationId: HomeworkController_getOne
      parameters:
        - name: homeworkId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a14
      tags: *a15
    patch:
      operationId: HomeworkController_update
      parameters:
        - name: homeworkId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HomeworkUpdateDto"
      responses:
        "200":
          description: ""
      security: *a14
      tags: *a15
    delete:
      operationId: HomeworkController_remove
      parameters:
        - name: homeworkId
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: ""
      security: *a14
      tags: *a15
...
info:
  title: Chess Coach Copilot
  description: ""
  version: 0.1.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateExternalAccountDto:
      type: object
      properties:
        platform:
          type: string
          enum:
            - LICHESS
            - CHESS_COM
          example: LICHESS
        username:
          type: string
          example: maksim_sokolov_2012
      required:
        - platform
        - username
    UpdateExternalAccountDto:
      type: object
      properties:
        platform:
          type: string
          enum:
            - LICHESS
            - CHESS_COM
          example: LICHESS
        username:
          type: string
          example: maksim_sokolov_2012
      required:
        - platform
        - username
    AnalysisCriticalMomentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        analysisId:
          type: string
        ply:
          type: number
        fullMoveNumber:
          type: number
        moveNumber:
          type: string
        moveColor:
          type: string
          enum:
            - WHITE
            - BLACK
        san:
          type: string
        lan:
          type: string
          nullable: true
        uci:
          type: string
          nullable: true
        beforeFen:
          type: string
        afterFen:
          type: string
        bestMove:
          type: string
          nullable: true
        bestVariation:
          type: array
          items:
            type: object
            additionalProperties: true
        nags:
          type: array
          items:
            type: object
            additionalProperties: true
        comments:
          type: array
          items:
            type: object
            additionalProperties: true
        evaluationBefore:
          type: object
          additionalProperties: true
          nullable: true
        evaluationAfter:
          type: object
          additionalProperties: true
          nullable: true
        severity:
          type: string
          enum:
            - INACCURACY
            - MISTAKE
            - BLUNDER
            - MATE
            - UNKNOWN
        sourceEvidence:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - analysisId
        - ply
        - fullMoveNumber
        - moveNumber
        - moveColor
        - san
        - lan
        - uci
        - beforeFen
        - afterFen
        - bestMove
        - bestVariation
        - nags
        - comments
        - evaluationBefore
        - evaluationAfter
        - severity
        - sourceEvidence
        - createdAt
        - updatedAt
    AnalysisMistakeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        analysisId:
          type: string
          format: uuid
        criticalMomentId:
          type: string
          format: uuid
          nullable: true
        severity:
          type: string
          enum:
            - INACCURACY
            - MISTAKE
            - BLUNDER
            - MATE
            - UNKNOWN
        category:
          type: string
        explanation:
          type: string
        suggestedFix:
          type: string
          nullable: true
        sourceEvidence:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - analysisId
        - criticalMomentId
        - severity
        - category
        - explanation
        - suggestedFix
        - sourceEvidence
        - createdAt
        - updatedAt
    AnalysisDetailsResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        analysisJobId:
          type: string
          format: uuid
        gameId:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        confidenceLevel:
          type: string
          enum:
            - HIGH
            - MEDIUM
            - LOW
        annotationCoverage:
          type: string
          enum:
            - NONE
            - PARTIAL
            - FULL
        reducedConfidenceWarning:
          type: string
          nullable: true
        overallDiagnosis:
          type: string
        openingName:
          type: string
          nullable: true
        result:
          type: string
          enum:
            - WIN
            - LOSS
            - DRAW
            - UNKNOWN
        mainWeaknessTag:
          type: string
          enum:
            - MISSED_FORK
            - MISSED_PIN
            - MISSED_DOUBLE_ATTACK
            - MISSED_DISCOVERED_ATTACK
            - MISSED_MATE
            - ALLOWED_MATE
            - HANGING_PIECE
            - MISSED_CAPTURE
            - BAD_CAPTURE
            - MISSED_OPPONENT_THREAT
            - OPENED_LINE
            - UNKNOWN_TACTICAL_ERROR
            - KING_SAFETY
            - DELAYED_CASTLING
            - POOR_DEVELOPMENT
            - EARLY_QUEEN
            - BAD_TRADE
            - POOR_CONVERSION
            - OPENING_STRATEGY
            - PAWN_STRUCTURE
            - ENDGAME_TECHNIQUE
            - UNKNOWN_STRATEGIC_ERROR
            - TIME_MANAGEMENT
            - CALCULATION_DEPTH
            - TUNNEL_VISION
            - MATERIAL_GREED
            - LOW_BOARD_AWARENESS
            - INSUFFICIENT_ANNOTATION_DATA
            - REDUCED_CONFIDENCE
          nullable: true
        secondaryWeaknessTags:
          type: array
          items:
            type: string
            enum:
              - MISSED_FORK
              - MISSED_PIN
              - MISSED_DOUBLE_ATTACK
              - MISSED_DISCOVERED_ATTACK
              - MISSED_MATE
              - ALLOWED_MATE
              - HANGING_PIECE
              - MISSED_CAPTURE
              - BAD_CAPTURE
              - MISSED_OPPONENT_THREAT
              - OPENED_LINE
              - UNKNOWN_TACTICAL_ERROR
              - KING_SAFETY
              - DELAYED_CASTLING
              - POOR_DEVELOPMENT
              - EARLY_QUEEN
              - BAD_TRADE
              - POOR_CONVERSION
              - OPENING_STRATEGY
              - PAWN_STRUCTURE
              - ENDGAME_TECHNIQUE
              - UNKNOWN_STRATEGIC_ERROR
              - TIME_MANAGEMENT
              - CALCULATION_DEPTH
              - TUNNEL_VISION
              - MATERIAL_GREED
              - LOW_BOARD_AWARENESS
              - INSUFFICIENT_ANNOTATION_DATA
              - REDUCED_CONFIDENCE
        recommendedLessonTitle:
          type: string
          nullable: true
        recommendedLessonWhy:
          type: string
          nullable: true
        recommendedFocusPoints:
          type: array
          items:
            type: string
        criticalMoments:
          type: array
          items:
            $ref: "#/components/schemas/AnalysisCriticalMomentResponse"
        mistakes:
          type: array
          items:
            $ref: "#/components/schemas/AnalysisMistakeResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - analysisJobId
        - gameId
        - studentId
        - confidenceLevel
        - annotationCoverage
        - reducedConfidenceWarning
        - overallDiagnosis
        - openingName
        - result
        - mainWeaknessTag
        - secondaryWeaknessTags
        - recommendedLessonTitle
        - recommendedLessonWhy
        - recommendedFocusPoints
        - criticalMoments
        - mistakes
        - createdAt
        - updatedAt
    ImportPgnDto:
      type: object
      properties:
        studentColor:
          type: string
          enum:
            - WHITE
            - BLACK
          example: WHITE
        sourceLabel:
          type: string
          example: Lichess Study export
        rawPgn:
          type: string
          example: >-
            [Event "Training Game"]

            [Site "Lichess"]

            [Date "2026.07.25"]

            [Round "?"]

            [White "Student"]

            [Black "Opponent"]

            [Result "1-0"]


            1. e4 { [%eval 0.3] Good central control. } e5 { [%eval 0.1] } 2.
            Nf3 { [%eval 0.5] } Nc6 { [%eval 0.2] } 3. Bb5 { [%eval 0.9] } a6 {
            [%eval 0.4] } 1-0
      required:
        - studentColor
        - rawPgn
...
    HomeworkUpdateDto:
      type: object
      properties:
        title:
          type: string
          example: "Homework: candidate move discipline"
        content:
          type: object
          example:
            overview: Review the forcing-move checklist before every capture.
            exercises:
              - Annotate move 18 alternatives
              - Solve 5 fork puzzles
            focusPoints:
              - Checks, captures, threats
            notes:
              - Discuss move-order discipline in next lesson
...